Main;Runtime: use size_t
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@33700 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ include/iprt/cpp/xml.h
@@ -635,7 +635,7 @@ public:
     XmlMemParser();
     ~XmlMemParser();
 
-    void read(const void* pvBuf, int cbSize, const iprt::MiniString &strFilename, Document &doc);
+    void read(const void* pvBuf, size_t cbSize, const iprt::MiniString &strFilename, Document &doc);
 };
 
 /*
